Refactor codebase
This should make it easier for me to maintain.
That's about it.
