+++ src/commands/disconnect.command.ts
@@ -1,6 +1,6 @@
+import { hasVoiceState } from '@app/utils';
 import { SlashCommandBuilder } from 'discord.js';
 import { AppCommand } from './command';
-import { hasVoiceState } from '@/utils/has-voice-state';
 
 export const disconnect: AppCommand = {
   data: new SlashCommandBuilder()
