+++ src/commands/loop.command.ts
@@ -1,6 +1,6 @@
+import { hasVoiceState } from '@app/utils/has-voice-state';
 import { SlashCommandBuilder, SlashCommandSubcommandBuilder } from 'discord.js';
 import { AppCommand } from './command';
-import { hasVoiceState } from '@/utils/has-voice-state';
 
 export const loop: AppCommand = {
   data: new SlashCommandBuilder()
