+++ src/commands/play.command.ts
@@ -1,24 +1,22 @@
-import { AppContext } from '@/app.context';
-import { AppEmoji } from '@/app.emojis';
-import { logger } from '@/logger';
-import { hasVoiceState } from '@/utils/has-voice-state';
-import { isSpotify } from '@/utils/is-spotify';
-import { trimEllipse } from '@/utils/trim-ellipses';
+import { AppContext } from '@app/context';
+import { logger } from '@app/logger';
+import { QueueType } from '@app/player';
+import { handleSearch, handleTrack, handleTracks } from '@app/player/handlers';
 import {
-  ActionRowBuilder,
+  handleSpotifyAlbum,
+  handleSpotifyPlaylist,
+  handleSpotifyTrack
+} from '@app/spotify/handlers';
+import { hasVoiceState, isSpotify, trimEllipse } from '@app/utils';
+import {
   ChatInputCommandInteraction,
-  ComponentType,
   SlashCommandBuilder,
   SlashCommandStringOption,
-  SlashCommandSubcommandBuilder,
-  StringSelectMenuBuilder,
-  StringSelectMenuOptionBuilder
+  SlashCommandSubcommandBuilder
 } from 'discord.js';
-import { MoonlinkPlayer, MoonlinkTrack } from 'moonlink.js';
+import { MoonlinkPlayer } from 'moonlink.js';
 import { parse as parseSpotifyUri } from 'spotify-uri';
 import { AppCommand } from './command';
-
-export type QueueType = 'later' | 'next' | 'now';
 
 export const play: AppCommand = {
   data: new SlashCommandBuilder()
@@ -139,8 +137,9 @@ const playMusic = async (options: {
     guildId: interaction.guild.id,
     voiceChannel: interaction.member.voice.channel.id,
     textChannel: interaction.channelId,
+    autoLeave: true,
     autoPlay: false,
-    volume: 100
+    loop: 0
   });
 
   if (player.voiceChannel !== interaction.member.voice.channel.id) {
@@ -151,10 +150,10 @@ const playMusic = async (options: {
     // Set text channel if not matching
     player.setTextChannel(interaction.channelId);
   }
-  if (!player.connected) {
-    // Connect to the voice channel if not connected
-    player.connect({ setMute: false, setDeaf: false });
-  }
+
+  logger.debug(`Connecting to voice channel: ${interaction.member.voice.channel.name}`);
+  // Connect to the voice channel if not connected
+  player.connect({ setMute: false, setDeaf: false });
 
   if (isSpotify(query)) {
     await handleSpotify({ query, player, context, interaction, queue });
@@ -162,7 +161,10 @@ const playMusic = async (options: {
     await handleYoutube({ query, player, context, interaction, queue });
   }
 
-  if (!player.playing && !player.paused && !player.current) {
+  if (player.paused) {
+    await player.resume();
+  }
+  if (!player.playing) {
     await player.play();
   }
 };
@@ -190,91 +192,33 @@ const handleYoutube = async (options: {
         ephemeral: true,
         content: `There was an error looking up the music. Please try again.`
       });
-      return;
+      break;
     }
     case 'empty': {
       // Responding with a message if the search returns no results
-      await interaction.followUp({
-        ephemeral: true,
-        content: `No matches found!`
-      });
+      await interaction.followUp({ ephemeral: true, content: `No matches found!` });
       return;
     }
     case 'playlist': {
-      if (queue !== 'later') {
-        await interaction.followUp({
-          ephemeral: true,
-          content: `Trying to load an entire playlist on priority is cheating.`
-        });
-        return;
-      }
-
       if (!result.playlistInfo) return;
-
-      for (const track of result.tracks) {
-        player.queue.add(track);
-      }
-
-      await interaction.followUp({
-        ephemeral: true,
-        content: `Queueing music list: \`${result.playlistInfo.name}\`.`
+      // Handle playlist result
+      await handleTracks({
+        interaction,
+        player,
+        queue,
+        tracks: result.tracks,
+        name: result.playlistInfo.name
       });
       break;
     }
     case 'track': {
-      await respondToPlay({ interaction, track: result.tracks[0], player, queue });
+      // Handle track result
+      await handleTrack({ interaction, track: result.tracks[0], player, queue });
       break;
     }
     case 'search': {
-      const selectMusicMenu = new StringSelectMenuBuilder()
-        .setCustomId(`select:music`)
-        .setPlaceholder('Please select music to play');
-
-      for (const [index, track] of result.tracks.entries()) {
-        selectMusicMenu.addOptions(
-          new StringSelectMenuOptionBuilder()
-            .setLabel(trimEllipse(track.title, 100))
-            .setDescription(trimEllipse(track.author, 100))
-            .setValue(track.identifier)
-            .setEmoji(index === 0 ? AppEmoji.Preferred : AppEmoji.MusicNote)
-        );
-      }
-
-      const row = new ActionRowBuilder<StringSelectMenuBuilder>().addComponents(selectMusicMenu);
-
-      const response = await interaction.followUp({
-        components: [row]
-      });
-
-      try {
-        const selectTrack = await response.awaitMessageComponent({
-          componentType: ComponentType.StringSelect,
-          filter: (i) => i.user.id === interaction.user.id,
-          time: 60_000
-        });
-        const selectedIdentifier = selectTrack.values[0];
-        const selectedTrack = result.tracks.find(
-          (track) => track.identifier === selectedIdentifier
-        );
-
-        if (!selectedTrack) {
-          await interaction.editReply({
-            content: `Unable to find selected track.`,
-            components: []
-          });
-          return;
-        }
-
-        logger.debug('Track selected', selectedTrack);
-
-        await respondToPlay({ interaction, track: selectedTrack, player, queue });
-      } catch (e) {
-        await interaction.editReply({
-          content: 'No music selected within a minute, cancelled.',
-          components: []
-        });
-        return;
-      }
+      // Handle search result
+      await handleSearch({ interaction, player, queue, result });
       break;
     }
   }
@@ -288,181 +232,22 @@ const handleSpotify = async (options: {
   queue: QueueType;
 }) => {
   const { query, interaction, queue, player, context } = options;
-  const { spotify } = context;
   const spotifyUri = parseSpotifyUri(query);
 
   switch (spotifyUri.type) {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
     case 'album': {
-      if (queue !== 'later') {
-        await interaction.followUp({
-          ephemeral: true,
-          content: `Trying to load an entire album on priority is cheating.`
-        });
-        return;
-      }
-
-      const spotifyAlbum = await spotify.albums.get(spotifyUri.id);
-
-      for (const spotifyTrack of spotifyAlbum.tracks.items) {
-        const spotifyArtists = spotifyTrack.artists.map((artist) => artist.name).join(' ');
-        const track = await lookupTrack({
-          query: `${spotifyTrack.name} ${spotifyArtists}`,
-          interaction,
-          context
-        });
-
-        if (track) {
-          player.queue.add(track);
-        }
-      }
-
-      await interaction.followUp({
-        ephemeral: true,
-        content: `Queueing spotify album \`${spotifyAlbum.name}\``
-      });
-
+      await handleSpotifyAlbum({ context, interaction, spotifyUri, player, queue });
       break;
     }
     // eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
     case 'track': {
-      const spotifyTrack = await spotify.tracks.get(spotifyUri.id);
-      const spotifyArtists = spotifyTrack.artists.map((artist) => artist.name).join(' ');
-      const track = await lookupTrack({
-        query: `${spotifyTrack.name} ${spotifyArtists}`,
-        interaction,
-        context
-      });
-
-      if (track) {
-        await respondToPlay({ interaction, track, player, queue });
-      } else {
-        await interaction.followUp({
-          ephemeral: true,
-          content: `There was an error looking up the music. Please try again.`
-        });
-      }
+      await handleSpotifyTrack({ context, interaction, spotifyUri, player, queue });
       break;
     }
     // eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
     case 'playlist': {
-      if (queue !== 'later') {
-        await interaction.followUp({
-          ephemeral: true,
-          content: `Trying to load an entire playlist on priority is cheating.`
-        });
-        return;
-      }
-
-      const playlistInfo = await spotify.playlists.getPlaylist(spotifyUri.id);
-
-      for (const playlistedSpotifyTrack of playlistInfo.tracks.items) {
-        const spotifyTrack = playlistedSpotifyTrack.track;
-        const spotifyArtists = spotifyTrack.artists.map((artist) => artist.name).join(' ');
-        const track = await lookupTrack({
-          query: `${spotifyTrack.name} ${spotifyArtists}`,
-          interaction,
-          context
-        });
-
-        if (track) {
-          player.queue.add(track);
-        }
-      }
-
-      await interaction.followUp({
-        ephemeral: true,
-        content: `Queueing spotify playlist \`${playlistInfo.name}\``
-      });
-
-      break;
-    }
-  }
-};
-
-const lookupTrack = async (options: {
-  query: string;
-  context: AppContext;
-  interaction: ChatInputCommandInteraction;
-}): Promise<MoonlinkTrack | null> => {
-  const { link } = options.context;
-  const { query, interaction } = options;
-  const result = await link.search({
-    query,
-    source: 'youtube',
-    requester: `<@${interaction.user.id}>`
-  });
-
-  switch (result.loadType) {
-    case 'error': {
-      logger.warn('Lookup error', { query });
-      return null;
-    }
-    case 'empty': {
-      logger.warn('Lookup returned', { query });
-      return null;
-    }
-    case 'search': {
-      return result.tracks[0];
-    }
-  }
-
-  return null;
-};
-
-const respondToPlay = async (options: {
-  interaction: ChatInputCommandInteraction;
-  track: MoonlinkTrack;
-  player: MoonlinkPlayer;
-  queue: QueueType;
-}) => {
-  const { interaction, track, player, queue } = options;
-
-  switch (queue) {
-    case 'later': {
-      player.queue.add(track);
-
-      if (!player.playing && !player.current) {
-        await interaction.editReply({
-          content: `Now playing \`${track.title}\`.`,
-          components: []
-        });
-      } else {
-        await interaction.editReply({
-          content: `Playing later \`${track.title}\`.`,
-          components: []
-        });
-      }
-      break;
-    }
-    case 'next': {
-      player.queue.add(track, 1);
-
-      if (!player.playing && !player.current) {
-        await interaction.editReply({
-          content: `Now playing \`${track.title}\`.`,
-          components: []
-        });
-      } else {
-        await interaction.editReply({
-          content: `Playing next \`${track.title}\`.`,
-          components: []
-        });
-      }
-      break;
-    }
-    case 'now': {
-      await player.play(track);
-
-      if (player.previous) {
-        const previousTrack = player.previous as MoonlinkTrack;
-        player.queue.add(previousTrack, 1);
-      }
-
-      await interaction.editReply({
-        content: `Now playing \`${track.title}\`.`,
-        components: []
-      });
+      await handleSpotifyPlaylist({ context, interaction, spotifyUri, player, queue });
       break;
     }
   }
