+++ src/emojis.ts
