+++ src/player/handlers/index.ts
@@ -1,0 +1,3 @@
+export * from './tracks.handler';
+export * from './search.handler';
+export * from './track.handler';
