+++ src/player/queue.type.ts
@@ -1,0 +1,1 @@
+export type QueueType = 'later' | 'next' | 'now';
