+++ src/spotify/handlers/index.ts
@@ -1,0 +1,3 @@
+export * from './playlist.handler';
+export * from './track.handler';
+export * from './album.handler';
