+++ tsconfig.json
@@ -10,7 +10,7 @@
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "paths": {
-      "@/*": ["./src/*"]
+      "@app/*": ["./src/*"]
     }
   },
   "include": ["src/**/*.ts"],
